Pages: プレビューのルートを分離
